Read defaults from Windows registry
This optionally enables an application to define a Windows
registry key in which defaults are looked up. This allows
KDE Software to be configured under Windows through Group
Policies.

Keys below this key are treated as configuration
groups. Values below this key are treated as values
of the default group.

To add a key to a group create a subkey under regKey and
add the value in this group with type REG_SZ or REG_EXPAND_SZ.

Immutable groups or values can be suffixed with [$i].

The values are read from HKLM and HKCU and treated as defaults.